Move print tag rendering into feature
